src/requirements.txt
# AI Agent Software Team POC - Dependencies

# Core LangGraph and LangChain
langgraph>=0.2.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-core>=0.3.0

# Qwen API
dashscope>=1.19.0

# Testing frameworks (for the testing agent's outputs)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# Code quality tools (optional, for enhancement)
black>=24.0.0
pylint>=3.0.0
mypy>=1.8.0

# For future integrations (commented out for POC)
# PyGithub>=2.1.1  # GitHub integration
# jira>=3.5.0  # Jira integration
# azure-devops>=7.1.0  # Azure DevOps integration

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.9.0
